testing common added, dnode updated with common mods
