+++ testing/services/dummy/dnode/config/config.go
@@ -8,49 +8,17 @@
 package config
 
 import (
-	"strconv"
-	"strings"
-
 	"github.com/prometheus/client_golang/prometheus"
+	cenums "github.com/ukama/ukama/testing/common/enums"
 )
 
-type AmqpConfig struct {
-	Uri      string `default:"http://rabbitmq:15672"`
-	Username string `default:"guest"`
-	Password string `default:"guest"`
-	Exchange string `default:"amq.topic"`
-	Vhost    string `default:"%2F"`
-}
-
-type Profile uint8
-
-const (
-	PROFILE_NORMAL Profile = 0
-	PROFILE_MIN    Profile = 1
-	PROFILE_MAX    Profile = 2
-)
-
-func ParseProfileType(value string) Profile {
-	i, err := strconv.Atoi(value)
-	if err == nil {
-		return Profile(i)
-	}
-
-	t := map[string]Profile{"normal": 0, "min": 1, "max": 2}
-
-	v, ok := t[strings.ToLower(value)]
-	if !ok {
-		return Profile(0)
-	}
-
-	return Profile(v)
-}
+const PORT = 8085
 
 type WMessage struct {
-	NodeId   string    `json:"nodeId"`
-	Profile  Profile   `json:"profile"`
-	Scenario SCENARIOS `json:"scenario"`
-	Kpis     NodeKPIs  `json:"kpis"`
+	Kpis     NodeKPIs         `json:"kpis"`
+	NodeId   string           `json:"nodeId"`
+	Profile  cenums.Profile   `json:"profile"`
+	Scenario cenums.SCENARIOS `json:"scenario"`
 }
 
 type NodeKPI struct {
@@ -64,8 +32,6 @@ type NodeKPI struct {
 type NodeKPIs struct {
 	KPIs []NodeKPI
 }
-
-const PORT = 8085
 
 var KPI_CONFIG = NodeKPIs{
 	KPIs: []NodeKPI{
@@ -251,35 +217,4 @@ var KPI_CONFIG = NodeKPIs{
 			),
 		},
 	},
-}
-
-type SCENARIOS string
-
-const (
-	SCENARIO_DEFAULT                SCENARIOS = "default"
-	SCENARIO_BACKHAUL_DOWN          SCENARIOS = "backhaul_down"
-	SCENARIO_BACKHAUL_DOWNLINK_DOWN SCENARIOS = "backhaul_downlink_down"
-	SCENARIO_SOLAR_DOWN             SCENARIOS = "solar_down"
-	SCENARIO_SWITCH_OFF             SCENARIOS = "switch_off"
-	SCENARIO_SITE_RESTART           SCENARIOS = "site_restart"
-	SCENARIO_NODE_OFF               SCENARIOS = "node_off"
-)
-
-func ParseScenarioType(value string) SCENARIOS {
-	t := map[string]SCENARIOS{
-		"default":                SCENARIO_DEFAULT,
-		"backhaul_down":          SCENARIO_BACKHAUL_DOWN,
-		"backhaul_downlink_down": SCENARIO_BACKHAUL_DOWNLINK_DOWN,
-		"solar_down":             SCENARIO_SOLAR_DOWN,
-		"switch_off":             SCENARIO_SWITCH_OFF,
-		"site_restart":           SCENARIO_SITE_RESTART,
-		"node_off":               SCENARIO_NODE_OFF,
-	}
-
-	v, ok := t[strings.ToLower(value)]
-	if !ok {
-		return SCENARIO_DEFAULT
-	}
-
-	return SCENARIOS(v)
 }
